requirements.txt
# ============================================
# 企业微信智能机器人平台 - 依赖包
# ============================================

# 加密和配置
pycryptodome==3.19.0
python-dotenv==1.0.0
pyyaml==6.0.1

# HTTP客户端
requests==2.31.0

# 异步HTTP客户端
httpx==0.28.1
aiohttp==3.9.1

# WebSocket客户端
websockets==14.2

# SSL证书
certifi>=2024.0

# AI 模型 SDK（可选，按需安装）
# Gemini API (使用 REST API，无需额外 SDK)

# OpenAI API
openai>=1.0.0
